// navoiy-ui/types/src/lib/pages.model.ts
export enum Pages {
  aboutUs = 'about-us',
  events = 'events',
  eventDetails = 'event-details',
  fullName = 'full-name',
  heads = 'heads',
  home = 'home',
  history = 'history',
  news = 'news',
  page404 = '404-page',
  newsDetails = 'news-details',
  team = 'team',
  page500 = '500-page',
}